+++ app/src/main/java/org/simple/clinic/drugs/selection/custom/CustomDrugEntryUiRenderer.kt
@@ -22,6 +22,8 @@ class CustomDrugEntryUiRenderer(
         setSheetTitle(model.drugName, model.dosage, model.drugFrequencyToFrequencyChoiceItemMap[model.frequency]!!.label)
 
       showDefaultDosagePlaceholder(model.dosage, model.dosageHasFocus)
+
+      if (model.isSaveButtonInProgressState) ui.showSaveButtonProgressState()
     }
   }