+++ app/src/test/java/org/simple/clinic/drugs/selection/custom/CustomDrugEntryUiRendererTest.kt
@@ -5,6 +5,7 @@ import com.nhaarman.mockitokotlin2.verify
 import com.nhaarman.mockitokotlin2.verifyNoMoreInteractions
 import org.junit.Test
 import org.simple.clinic.drugs.search.DrugFrequency
+import org.simple.clinic.drugs.selection.custom.ButtonState.SAVING
 import org.simple.clinic.drugs.selection.custom.drugfrequency.country.DrugFrequencyChoiceItem
 import java.util.UUID
 
@@ -23,7 +24,9 @@ class CustomDrugEntryUiRendererTest {
       DrugFrequency.QDS to DrugFrequencyChoiceItem(drugFrequency = DrugFrequency.QDS, label = "QDS")
   )
 
-  private val defaultModel = CustomDrugEntryModel.default(openAs = OpenAs.New.FromDrugName(drugName), dosagePlaceholder).drugFrequencyToFrequencyChoiceItemMapLoaded(drugFrequencyToFrequencyChoiceItemMap)
+  private val defaultModel = CustomDrugEntryModel
+      .default(openAs = OpenAs.New.FromDrugName(drugName), dosagePlaceholder)
+      .drugFrequencyToFrequencyChoiceItemMapLoaded(drugFrequencyToFrequencyChoiceItemMap)
 
   @Test
   fun `when drug dosage focus is changed and dosage is null, then set drug dosage text with the placeholder and move the cursor to the beginning`() {
@@ -52,7 +55,11 @@ class CustomDrugEntryUiRendererTest {
   fun `when drug dosage focus is changed and dosage is not null but only contains the placeholder, then set drug dosage text as an empty string and update the sheet title`() {
     // given
     val dosageText = "mg"
-    val drugDosageChangedModel = defaultModel.drugNameLoaded(drugName).dosageEdited(dosageText).dosageFocusChanged(hasFocus = false).drugInfoProgressStateLoaded()
+    val drugDosageChangedModel = defaultModel
+        .drugNameLoaded(drugName)
+        .dosageEdited(dosageText)
+        .dosageFocusChanged(hasFocus = false)
+        .drugInfoProgressStateLoaded()
     val frequencyLabel = "None"
 
     // when
@@ -73,7 +80,10 @@ class CustomDrugEntryUiRendererTest {
   fun `when the screen is loaded in update mode, then render the drug name and setup ui for updating drug entry`() {
     // given
     val prescribedDrugUuid = UUID.fromString("96633994-6e4d-4528-b796-f03ae016553a")
-    val defaultModel = CustomDrugEntryModel.default(openAs = OpenAs.Update(prescribedDrugUuid), dosagePlaceholder).drugFrequencyToFrequencyChoiceItemMapLoaded(drugFrequencyToFrequencyChoiceItemMap).drugInfoProgressStateLoaded()
+    val defaultModel = CustomDrugEntryModel
+        .default(openAs = OpenAs.Update(prescribedDrugUuid), dosagePlaceholder)
+        .drugFrequencyToFrequencyChoiceItemMapLoaded(drugFrequencyToFrequencyChoiceItemMap)
+        .drugInfoProgressStateLoaded()
     val frequencyLabel = "None"
 
     // when
@@ -134,6 +144,26 @@ class CustomDrugEntryUiRendererTest {
     verify(ui).setButtonTextAsAdd()
     verify(ui).showKeyboard()
     verify(ui).setSheetTitle(null, null, frequencyLabel)
+    verifyNoMoreInteractions(ui)
+  }
+
+
+  @Test
+  fun `when add button is clicked and info is being added or updated, then show progress state in the save button`() {
+    // given
+    val frequencyLabel = "None"
+
+    // when
+    uiRenderer.render(defaultModel.drugInfoProgressStateLoaded().saveButtonStateChanged(SAVING))
+
+    // then
+    verify(ui).hideProgressBar()
+    verify(ui).showCustomDrugEntryUi()
+    verify(ui).hideRemoveButton()
+    verify(ui).setButtonTextAsAdd()
+    verify(ui).showKeyboard()
+    verify(ui).setSheetTitle(null, null, frequencyLabel)
+    verify(ui).showSaveButtonProgressState()
     verifyNoMoreInteractions(ui)
   }
 }